requirements.txt
asyncio
hikari
hikari-lightbulb
apscheduler
uvloop
nekos.py
git+https://github.com/neonjonn/lightbulb-ext-neon.git
